[None][chore] Merge spec_decode ops into ops/flashinfer
The spec-decoding samplers moved into flashinfer.py (they hard-depend on
flashinfer and combine it with vanilla's temperature helpers), removing the
separate spec_decode.py. Speculative call sites import from ops.flashinfer /
ops.vanilla directly.
